+++ pytsite/odm/_error.py
@@ -21,10 +21,6 @@ class EntityNotStored(Exception):
     pass
 
 
-class ForbidEntityOperation(Exception):
-    pass
-
-
 class FieldNotDefined(Exception):
     pass